requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-Script==2.0.6
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
Flask-Admin==1.5.6
Flask-Babelex==0.9.4
email_validator==1.0.5
sqlalchemy==1.3.15
sqlalchemy-utils==0.36.3
python-dotenv==0.12.0
requests==2.23.0
gunicorn==20.0.4
psycopg2==2.8.5
psycopg2-binary==2.8.5